+++ darts/tests/dataprocessing/transformers/test_invertible_fittable_data_transformer.py
@@ -136,7 +136,10 @@ class TestLocalFittableInvertibleDataTransformer:
 
         @staticmethod
         def ts_inverse_transform(
-            series: TimeSeries, params: Mapping[str, Any], **kwargs
+            series: TimeSeries,
+            params: Mapping[str, Any],
+            insample: TimeSeries | None = None,
+            **kwargs,
         ) -> TimeSeries:
             """
             Implements the inverse transform `(series - translation) / scale`.
@@ -491,6 +494,32 @@ class TestLocalFittableInvertibleDataTransformer:
         inv = mock.inverse_transform(transformed, component_mask=mask)
         assert inv == test_input
 
+    def test_inverse_tf_on_nested_series_with_series_idx(self):
+        """
+        Tests for correct (inverse) transformation for nested lists of input series with `series_idx`.
+        """
+        inp_1 = constant_timeseries(value=1, length=10)
+        inp_2 = constant_timeseries(value=2, length=10)
+
+        mock = self.DataTransformerMock(scale=2, translation=10)
+
+        tf_1, tf_2 = mock.fit_transform([inp_1, inp_2])
+
+        nested_inp = [[tf_1[1:], tf_1[2:]], [tf_2[1:]]]
+        inv_tf_direct = mock.inverse_transform(nested_inp)
+        inv_tf_direct_reversed = mock.inverse_transform(
+            nested_inp[::-1], series_idx=[1, 0]
+        )[::-1]
+        inv_tf_separate = [
+            mock.inverse_transform([nested_inp[idx]], series_idx=idx)[0]
+            for idx in range(2)
+        ]
+        assert inv_tf_direct == inv_tf_direct_reversed
+        assert inv_tf_separate == inv_tf_direct
+        assert inv_tf_separate[0][0] == inp_1[1:]
+        assert inv_tf_separate[0][1] == inp_1[2:]
+        assert inv_tf_separate[1][0] == inp_2[1:]
+
 
 class TestGlobalFittableInvertibleDataTransformer:
     """
@@ -560,11 +589,12 @@ class TestGlobalFittableInvertibleDataTransformer:
 
         @staticmethod
         def ts_inverse_transform(
-            series: TimeSeries, params: Mapping[str, Any], **kwargs
+            series: TimeSeries,
+            params: Mapping[str, Any],
+            insample: TimeSeries | None = None,
+            **kwargs,
         ) -> TimeSeries:
-            """
-            Implements the inverse transform `series + mean`.
-            """
+            """Implements the inverse transform `series + mean`."""
             mean = params["fitted"]
             vals = series.all_values()
             vals += mean